# src/main/java/com/probestack/forgestudio/design/api/DocumentsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CreateDocumentRequest;
import com.probestack.forgestudio.design.model.Document;
import com.probestack.forgestudio.design.model.DocumentListResponse;
import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.UpdateDocumentRequest;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-18T02:37:18.740676733Z[GMT]")@Validated
@Tag(name = "Documents", description = "Operations on stored documents")
public interface DocumentsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /document : Register a new document
     * Creates a document record and returns a pre-signed upload URL. The &#x60;checksum&#x60; must be unique per owner.
     *
     * @param createDocumentRequest  (required)
     * @return Document registered successfully (status code 201)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or A document with this checksum already exists (status code 409)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "createDocument",
        summary = "Register a new document",
        description = "Creates a document record and returns a pre-signed upload URL. The `checksum` must be unique per owner.",
        tags = { "Documents" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Document registered successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Document.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "409", description = "A document with this checksum already exists", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/document",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Document> createDocument(
        @Parameter(name = "CreateDocumentRequest", description = "", required = true) @Valid @RequestBody CreateDocumentRequest createDocumentRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"fileName\" : \"supplier-contract-2026.pdf\", \"retentionUntil\" : \"2031-08-12T00:00:00.000+00:00\", \"displayName\" : \"Supplier Contract 2026\", \"downloadUrl\" : \"https://storage.example.com/signed/d0c1u2m3?exp=1786574059\", \"legalHold\" : false, \"ownerId\" : \"team-legal\", \"version\" : 3, \"sizeBytes\" : 481203, \"tags\" : [ \"contract\", \"supplier\", \"2026\" ], \"checksum\" : \"9f86d081884c7d659a2feaa0c55ad015\", \"uploadedAt\" : \"2026-08-12T10:00:00Z\", \"id\" : \"d0c1u2m3\", \"contentType\" : \"application/pdf\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * DELETE /documents/{documentId} : Delete a document
     * Permanently removes the document and all of its versions. Documents under legal hold cannot be deleted.
     *
     * @param documentId Unique document identifier (UUID) (required)
     * @return Document deleted successfully (no content returned) (status code 204)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "deleteDocument",
        summary = "Delete a document",
        description = "Permanently removes the document and all of its versions. Documents under legal hold cannot be deleted.",
        tags = { "Documents" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Document deleted successfully (no content returned)"),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/documents/{documentId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Void> deleteDocument(
        @Parameter(name = "documentId", description = "Unique document identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("documentId") String documentId
    ) {
                        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                    }


    /**
     * GET /documents/{documentId} : Get a document by ID
     * Retrieves document metadata and a short-lived signed download URL.
     *
     * @param documentId Unique document identifier (UUID) (required)
     * @return Document found (status code 200)
     *         or Document not found (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "getDocumentById",
        summary = "Get a document by ID",
        description = "Retrieves document metadata and a short-lived signed download URL.",
        tags = { "Documents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Document found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Document.class))
            }),
            @ApiResponse(responseCode = "404", description = "Document not found", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/documents/{documentId}",
        produces = { "application/json" }
    )
    
    default ResponseEntity<Document> getDocumentById(
        @Parameter(name = "documentId", description = "Unique document identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("documentId") String documentId
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"fileName\" : \"supplier-contract-2026.pdf\", \"retentionUntil\" : \"2031-08-12T00:00:00.000+00:00\", \"displayName\" : \"Supplier Contract 2026\", \"downloadUrl\" : \"https://storage.example.com/signed/d0c1u2m3?exp=1786574059\", \"legalHold\" : false, \"ownerId\" : \"team-legal\", \"version\" : 3, \"sizeBytes\" : 481203, \"tags\" : [ \"contract\", \"supplier\", \"2026\" ], \"checksum\" : \"9f86d081884c7d659a2feaa0c55ad015\", \"uploadedAt\" : \"2026-08-12T10:00:00Z\", \"id\" : \"d0c1u2m3\", \"contentType\" : \"application/pdf\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /document : List all documents
     * Returns a paginated list of documents, optionally filtered by owner or content type.
     *
     * @param ownerId Filter by the owning user or team (optional)
     * @param contentType Filter by MIME type (e.g., \&quot;application/pdf\&quot;) (optional)
     * @param page Page number (starts at 1) (optional, default to 1)
     * @param size Number of items per page (optional, default to 20)
     * @return Successful response (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "listDocuments",
        summary = "List all documents",
        description = "Returns a paginated list of documents, optionally filtered by owner or content type.",
        tags = { "Documents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Successful response", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = DocumentListResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/document",
        produces = { "application/json" }
    )
    
    default ResponseEntity<DocumentListResponse> listDocuments(
        @Parameter(name = "ownerId", description = "Filter by the owning user or team", in = ParameterIn.QUERY) @Valid @RequestParam(value = "ownerId", required = false) String ownerId,
        @Parameter(name = "contentType", description = "Filter by MIME type (e.g., \"application/pdf\")", in = ParameterIn.QUERY) @Valid @RequestParam(value = "contentType", required = false) String contentType,
        @Min(1) @Parameter(name = "page", description = "Page number (starts at 1)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "1") Integer page,
        @Min(1) @Max(100) @Parameter(name = "size", description = "Number of items per page", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"size\" : 20, \"totalPages\" : 47, \"page\" : 1, \"content\" : [ { \"fileName\" : \"supplier-contract-2026.pdf\", \"retentionUntil\" : \"2031-08-12T00:00:00.000+00:00\", \"displayName\" : \"Supplier Contract 2026\", \"downloadUrl\" : \"https://storage.example.com/signed/d0c1u2m3?exp=1786574059\", \"legalHold\" : false, \"ownerId\" : \"team-legal\", \"version\" : 3, \"sizeBytes\" : 481203, \"tags\" : [ \"contract\", \"supplier\", \"2026\" ], \"checksum\" : \"9f86d081884c7d659a2feaa0c55ad015\", \"uploadedAt\" : \"2026-08-12T10:00:00Z\", \"id\" : \"d0c1u2m3\", \"contentType\" : \"application/pdf\" }, { \"fileName\" : \"supplier-contract-2026.pdf\", \"retentionUntil\" : \"2031-08-12T00:00:00.000+00:00\", \"displayName\" : \"Supplier Contract 2026\", \"downloadUrl\" : \"https://storage.example.com/signed/d0c1u2m3?exp=1786574059\", \"legalHold\" : false, \"ownerId\" : \"team-legal\", \"version\" : 3, \"sizeBytes\" : 481203, \"tags\" : [ \"contract\", \"supplier\", \"2026\" ], \"checksum\" : \"9f86d081884c7d659a2feaa0c55ad015\", \"uploadedAt\" : \"2026-08-12T10:00:00Z\", \"id\" : \"d0c1u2m3\", \"contentType\" : \"application/pdf\" } ], \"totalElements\" : 940 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * PUT /documents/{documentId} : Update document metadata
     * Updates the display name, tags, or retention policy. The stored bytes and &#x60;checksum&#x60; cannot be changed.
     *
     * @param documentId Unique document identifier (UUID) (required)
     * @param updateDocumentRequest  (required)
     * @return Document updated successfully (status code 200)
     *         or Invalid request - the server could not understand the payload or parameters. (status code 400)
     *         or The requested resource was not found. (status code 404)
     *         or Internal server error - something went wrong on the server side. (status code 500)
     */
    @Operation(
        operationId = "updateDocument",
        summary = "Update document metadata",
        description = "Updates the display name, tags, or retention policy. The stored bytes and `checksum` cannot be changed.",
        tags = { "Documents" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Document updated successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Document.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request - the server could not understand the payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource was not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error - something went wrong on the server side.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/documents/{documentId}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<Document> updateDocument(
        @Parameter(name = "documentId", description = "Unique document identifier (UUID)", required = true, in = ParameterIn.PATH) @PathVariable("documentId") String documentId,
        @Parameter(name = "UpdateDocumentRequest", description = "", required = true) @Valid @RequestBody UpdateDocumentRequest updateDocumentRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"fileName\" : \"supplier-contract-2026.pdf\", \"retentionUntil\" : \"2031-08-12T00:00:00.000+00:00\", \"displayName\" : \"Supplier Contract 2026\", \"downloadUrl\" : \"https://storage.example.com/signed/d0c1u2m3?exp=1786574059\", \"legalHold\" : false, \"ownerId\" : \"team-legal\", \"version\" : 3, \"sizeBytes\" : 481203, \"tags\" : [ \"contract\", \"supplier\", \"2026\" ], \"checksum\" : \"9f86d081884c7d659a2feaa0c55ad015\", \"uploadedAt\" : \"2026-08-12T10:00:00Z\", \"id\" : \"d0c1u2m3\", \"contentType\" : \"application/pdf\" }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}